requirements.txt
torch
torch-geometric
transformers
rdkit
fastapi
uvicorn
pandas
scikit-learn
matplotlib
pytdc